Gauge convert and fix issues
